Fix failing python tests in jenkins.

+++ le-dataplatform/src/test/python/algorithmtest/dt_traintest.py
@@ -1,4 +1,5 @@
 import algorithmtestbase as at
+import os
 import shutil
 import unittest
 
@@ -6,7 +7,8 @@ class LogisticRegressionTest(unittest.TestCase, at.AlgorithmTestBase):
 
     @classmethod
     def setUpClass(cls):
-        shutil.rmtree("./results")
+        if os.path.exists("./results"):
+            shutil.rmtree("./results")
 
     def testTrain(self):
         algorithmProperties = { "criterion":"entropy" }
